Toma de contacto

+++ src/algo.java
@@ -1,8 +1,8 @@
 
 public class algo {
 
-	public static void f(int a) {
-		int n=a;
+	public static void f(long a) {
+		long n=a;
 		while(n>0) {
 			n--;
 		}
